- Replaced SharedBuffersCore with a proxy implementation for buffers and streams - Added a new function getFragment() for accessing buffer memory directly.
This is an attempt to find a sweet spot between efficiency,
genericity and ease of use.

getFragment() can be implemented by any kind of buffer with little
overhead, so it is in fact completely generic. It is also quite
efficient since it allows direct access to the buffer memory without
the extra copy that is required when using the read() function. Ease
of use - we'll see :)